+++ FogCarport/src/main/java/presentation/commands/Link.java
@@ -1,8 +1,9 @@
 /*
  * 
  */
-package presentation;
+package presentation.commands;
 
+import presentation.Validation;
 import data.exceptions.UserException;
 import javax.servlet.http.HttpServletRequest;
 import logic.LogicFacade;
@@ -14,7 +15,7 @@ import logic.LogicFacade;
 public class Link extends Command
 {
     @Override
-    String execute(HttpServletRequest request, LogicFacade logic) throws UserException
+    public String execute(HttpServletRequest request, LogicFacade logic) throws UserException
     {
         Validation validation = new Validation();
         return validation.validateString(request.getParameter("link"), "Link");
