+++ FogCarport/src/main/java/presentation/commands/MakeCarportForm.java
@@ -1,4 +1,4 @@
-package presentation;
+package presentation.commands;
 
 import data.exceptions.DataException;
 import data.exceptions.UserException;
@@ -13,7 +13,7 @@ import logic.LogicFacade;
 public class MakeCarportForm extends Command
 {
     @Override
-    String execute(HttpServletRequest request, LogicFacade logic) throws UserException, DataException
+    public String execute(HttpServletRequest request, LogicFacade logic) throws UserException, DataException
     {
         /* Initializing required variables */
