+++ FogCarport/src/main/java/presentation/commands/UnknownCommand.java
@@ -1,4 +1,4 @@
-package presentation;
+package presentation.commands;
 
 import data.exceptions.UserException;
 import javax.servlet.http.HttpServletRequest;
@@ -10,7 +10,7 @@ import logic.LogicFacade;
 class UnknownCommand extends Command
 {
     @Override
-    String execute(HttpServletRequest request, LogicFacade logic) throws UserException
+    public String execute(HttpServletRequest request, LogicFacade logic) throws UserException
     {
         if (request.getAttribute("target") == null){
             return "login";
